% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly.R
\name{ois_layout}
\alias{ois_layout}
\title{Plotly OIS styling function}
\usage{
ois_layout(fig)
}
\arguments{
\item{fig}{Plotly figure to style}
}
\description{
This function applies OIS styling to your plotly plot.
}
\examples{
ois_layout()
}
\keyword{OIS}
\keyword{style}
